move skeleton quality test to skeleton_test.py
